# perf: minimize identifier lookahead when parsing let