--- a/packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-escape-id/output.json
+++ b/packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-escape-id/output.json
@@ -1,0 +1,30 @@
+{
+  "type": "File",
+  "start":0,"end":13,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":13}},
+  "program": {
+    "type": "Program",
+    "start":0,"end":13,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":13}},
+    "sourceType": "script",
+    "interpreter": null,
+    "body": [
+      {
+        "type": "VariableDeclaration",
+        "start":0,"end":13,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":13}},
+        "declarations": [
+          {
+            "type": "VariableDeclarator",
+            "start":4,"end":12,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":12}},
+            "id": {
+              "type": "Identifier",
+              "start":4,"end":12,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":12},"identifierName":"ina"},
+              "name": "ina"
+            },
+            "init": null
+          }
+        ],
+        "kind": "let"
+      }
+    ],
+    "directives": []
+  }
+}