--- a/packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-non-BMP/output.json
+++ b/packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-non-BMP/output.json
@@ -1,0 +1,30 @@
+{
+  "type": "File",
+  "start":0,"end":23,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":23}},
+  "program": {
+    "type": "Program",
+    "start":0,"end":23,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":23}},
+    "sourceType": "script",
+    "interpreter": null,
+    "body": [
+      {
+        "type": "VariableDeclaration",
+        "start":0,"end":23,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":23}},
+        "declarations": [
+          {
+            "type": "VariableDeclarator",
+            "start":4,"end":22,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":22}},
+            "id": {
+              "type": "Identifier",
+              "start":4,"end":22,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":22},"identifierName":"in𝐬𝐭𝐚𝐧𝐜𝐞𝐨𝐟"},
+              "name": "in𝐬𝐭𝐚𝐧𝐜𝐞𝐨𝐟"
+            },
+            "init": null
+          }
+        ],
+        "kind": "let"
+      }
+    ],
+    "directives": []
+  }
+}