--- a/packages/babel-parser/test/fixtures/es2015/let/let-declaration-non-BMP-identifier/output.json
+++ b/packages/babel-parser/test/fixtures/es2015/let/let-declaration-non-BMP-identifier/output.json
@@ -1,0 +1,30 @@
+{
+  "type": "File",
+  "start":0,"end":9,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":9}},
+  "program": {
+    "type": "Program",
+    "start":0,"end":9,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":9}},
+    "sourceType": "script",
+    "interpreter": null,
+    "body": [
+      {
+        "type": "VariableDeclaration",
+        "start":0,"end":9,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":9}},
+        "declarations": [
+          {
+            "type": "VariableDeclarator",
+            "start":4,"end":8,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":8}},
+            "id": {
+              "type": "Identifier",
+              "start":4,"end":8,"loc":{"start":{"line":1,"column":4},"end":{"line":1,"column":8},"identifierName":"𝐢𝐧"},
+              "name": "𝐢𝐧"
+            },
+            "init": null
+          }
+        ],
+        "kind": "let"
+      }
+    ],
+    "directives": []
+  }
+}