… packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-escape-id/input.js
… packages/babel-parser/test/fixtures/es2015/let/let-declaration-in-non-BMP/input.js
… packages/babel-parser/test/fixtures/es2015/let/let-declaration-non-BMP-identifier/input.js