<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
When reviewing #13325 I realize that we always peek a complete identifier after `let` in order to compare with `in` and `instanceof`. However, we could exit early if the next character is an IdentfierStart and from this position on it does not match `in` or `instanceof`.

In a moderate `let` + length-10 identifier (not starting with `i`) benchmark cases, I see 17% performance gain. This benchmark is included in this PR.
```
node --predictable ./10-length-binding.mjs
baseline 64 let and length-10 binding identifiers: 5257 ops/sec ±45.3% (0.19ms)
baseline 128 let and length-10 binding identifiers: 3482 ops/sec ±0.28% (0.287ms)
baseline 256 let and length-10 binding identifiers: 1712 ops/sec ±0.75% (0.584ms)
baseline 512 let and length-10 binding identifiers: 848 ops/sec ±0.34% (1.179ms)
current 64 let and length-10 binding identifiers: 6109 ops/sec ±49.27% (0.164ms)
current 128 let and length-10 binding identifiers: 4018 ops/sec ±0.93% (0.249ms)
current 256 let and length-10 binding identifiers: 2005 ops/sec ±0.36% (0.499ms)
current 512 let and length-10 binding identifiers: 966 ops/sec ±0.89% (1.036ms)
```

If we go to extreme and benchmark with length-100 identifier, the performance gain is 30%. This is expected since it completely avoid lookahead 100 characters to compare with `in` or `instanceof`. The benchmark is not included in this PR.
```
baseline 64 let and length-100 binding identifiers: 2934 ops/sec ±45.04% (0.341ms)
baseline 128 let and length-100 binding identifiers: 2036 ops/sec ±1.22% (0.491ms)
baseline 256 let and length-100 binding identifiers: 1003 ops/sec ±1.21% (0.997ms)
baseline 512 let and length-100 binding identifiers: 484 ops/sec ±0.95% (2.067ms)
current 64 let and length-100 binding identifiers: 4252 ops/sec ±34.14% (0.235ms)
current 128 let and length-100 binding identifiers: 2636 ops/sec ±1.15% (0.379ms)
current 256 let and length-100 binding identifiers: 1318 ops/sec ±1.4% (0.759ms)
current 512 let and length-100 binding identifiers: 658 ops/sec ±0.54% (1.52ms)
```

This PR also fixed handling of non-BMP and escaped identifiers after `let`, the following valid cases are now parsed successfully:

```js
let in\u0061;
let in𝐬𝐭𝐚𝐧𝐜𝐞𝐨𝐟;
let 𝐢𝐧;
```

<a href="https://gitpod.io/#https://github.com/babel/babel/pull/13328"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>